vms/server/plugins/analytics/mqtt_plugin/python_src/requirements.txt
requests==2.32.3
paho-mqtt==2.1.0
aiohttp==3.10.10
